clean up portage Dockerfile
Having these commands in one layer does not decrease the image size and
splitting them up increases understanding of the Dockerfile.

Portage should not inherit from any other overlays and should not have a
masters = gentoo line.  The default layout.conf values are the desired
values since this image is meant to be the gentoo repository in
containers that want to use its volume.